# D3vz API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "D3vz API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/API/Aula": {
      "post": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "profId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "alunoId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "datahora",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "url",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "datahora",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "url",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/Aula/FromBody": {
      "post": {
        "tags": [
          "Aula"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JAula"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Aula"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JAula"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/Aula/GetByUser": {
      "get": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "discriminacao",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/Aula/GetByInterlocutores": {
      "get": {
        "tags": [
          "Aula"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "professorId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "alunoId",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/InterQuali": {
      "delete": {
        "tags": [
          "InterQuali"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "InterQuali"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "linguagem",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/InterQuali/GetById": {
      "get": {
        "tags": [
          "InterQuali"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/InterQuali/GetAll": {
      "get": {
        "tags": [
          "InterQuali"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/User": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "Id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          },
          {
            "in": "query",
            "name": "discriminacao",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/User/GetById": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/User/GetAll": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "discriminacao",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/User/Auth": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/API/User/AuthGoogle": {
      "post": {
        "tags": [
          "User"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/JUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "JAula": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "alunoid": {
          "format": "int64",
          "type": "integer"
        },
        "profid": {
          "format": "int64",
          "type": "integer"
        },
        "datahora": {
          "format": "date-time",
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "tempo": {
          "format": "int32",
          "type": "integer"
        },
        "aceito": {
          "type": "boolean"
        }
      }
    },
    "JUser": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "discriminacao": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "descricao": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "nascimento": {
          "format": "date-time",
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "interquali": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "foto": {
          "type": "string"
        },
        "dias": {
          "type": "string"
        },
        "horarios": {
          "type": "string"
        },
        "googleauth": {
          "type": "string"
        }
      }
    }
  }
}